Version: Change function signature to signed

+++ python/version.cc
@@ -8,18 +8,18 @@
 
 namespace bp = boost::python;
 
-inline bool checkVersionAtLeast(unsigned int major,
-                                unsigned int minor,
-                                unsigned int patch)
+inline bool checkVersionAtLeast(int major,
+                                int minor,
+                                int patch)
 {
-  return HPP_FCL_VERSION_AT_LEAST((int)major, (int)minor, (int)patch);
+  return HPP_FCL_VERSION_AT_LEAST(major, minor, patch);
 }
 
-inline bool checkVersionAtMost(unsigned int major,
-                               unsigned int minor,
-                               unsigned int patch)
+inline bool checkVersionAtMost(int major,
+                               int minor,
+                               int patch)
 {
-  return HPP_FCL_VERSION_AT_MOST((int)major, (int)minor, (int)patch);
+  return HPP_FCL_VERSION_AT_MOST(major, minor, patch);
 }
 
 void exposeVersion()
